+++ jsx/src/AppEx6.js
@@ -1,27 +1,5 @@
 import React from "react";
-import styled, { css } from "styled-components";
-
-const ButtonRound = css`
-  border-radius: 10px;
-  border: 0;
-  box-shadow: 2px 2px 2px rgba(0, 0, 0, 0.2);
-`;
-
-const Button1 = styled.button`
-  background-color: royalblue;
-  padding: 0.5em;
-  color: white;
-`;
-
-const Button2 = styled(Button1)`
-  ${ButtonRound}
-  color: black;
-`;
-
-const Button3 = styled(Button1)`
-  ${ButtonRound}
-  background-color: greenyellow;
-`;
+import { Button1, Button2, Button3 } from "./components/Buttons";
 
 export default function AppEx6() {
   return (
